Improve regular expressions
This commit improves the regular expressions parsing the meal
information from the mensa urls. The number of regular expressions was
reduced and further information is extracted, namely the different
prices for different people and the allergens (currently represented as
a list of abbreviations). These fields have also been added to the data
class Essen for future use.